Support alphabetical sort in filtered tree
