+++ src/stores/nodeDefStore.ts
@@ -226,27 +226,6 @@ export const SYSTEM_NODE_DEFS: Record<string, ComfyNodeDef> = {
   }
 }
 
-function sortedTree(node: TreeNode): TreeNode {
-  // Create a new node with the same label and data
-  const newNode: TreeNode = {
-    ...node
-  }
-
-  if (node.children) {
-    // Sort the children of the current node
-    const sortedChildren = [...node.children].sort((a, b) =>
-      a.label.localeCompare(b.label)
-    )
-    // Recursively sort the children and add them to the new node
-    newNode.children = []
-    for (const child of sortedChildren) {
-      newNode.children.push(sortedTree(child))
-    }
-  }
-
-  return newNode
-}
-
 interface State {
   nodeDefsByName: Record<string, ComfyNodeDefImpl>
   widgets: Record<string, ComfyWidgetConstructor>
@@ -269,9 +248,6 @@ export const useNodeDefStore = defineStore('nodeDef', {
         ...nodeDef.category.split('/'),
         nodeDef.display_name
       ])
-    },
-    sortedNodeTree(): TreeNode {
-      return sortedTree(this.nodeTree)
     }
   },
   actions: {
