+++ src/utils/treeUtil.ts
@@ -47,3 +47,24 @@ export function flattenTree<T>(tree: TreeNode): T[] {
   }
   return result
 }
+
+export function sortedTree(node: TreeNode): TreeNode {
+  // Create a new node with the same label and data
+  const newNode: TreeNode = {
+    ...node
+  }
+
+  if (node.children) {
+    // Sort the children of the current node
+    const sortedChildren = [...node.children].sort((a, b) =>
+      a.label.localeCompare(b.label)
+    )
+    // Recursively sort the children and add them to the new node
+    newNode.children = []
+    for (const child of sortedChildren) {
+      newNode.children.push(sortedTree(child))
+    }
+  }
+
+  return newNode
+}
